// YOUCOM.ReserVook/src/app/features/webreserves/model/webreservebase.model.ts
export class WebReserveBaseInfo {
  companyNo: string;
  scCd: string;
  scRcvSeq: number;

  xDataClsfic: string;
  xDataId: string;

  xSalesOfcCompanyCd: string;
  xSalesOfcCompanyNm: string;
  xSalesOfcNm: string;
  xSalesOfcCd: string;
  xSalesOfcPrsnInChg: string;
  xSalesOfcEmail: string;
  xSalesOfcPhnNum: string;

  xTravelAgncBkngNum: string;
  xTravelAgncBkngDate: string;
  xTravelAgncBkngTime: string;
  xTravelAgncReportNum: string;

  xGstOrGpNmSnglBt: string;
  xGstOrGpNmDoubleBt: string;
  xGstOrGpNmKanjiNm: string;

  xCheckInDate: string;
  xCheckInTime: string;
  xCheckOutDate: string;
  xCheckOutTime: string;
  xNights: number;

  xTtlRmCnt: number;
  xGrandTtlPaxCnt: number;

  xPackagePlanNm: string;
  xPackagePlanCd: string;

  reservationNo: string;
  scProcessedCd: string;
  scProcessedMessage: string;
  updateDatetime: string;

  type: string;
  typeColor: string;

  newCount:  number;
  cancelCount:  number;
}